Implemented player clamp to board, see details
The player is checked for illegal position serverside. If player enter illegal space--> warning is thrown and the player position is corrected by the server by sending new coordinates that the players position is set to. So no local client code evaluate if the position is illegal or not. Some clean up